remove metrics
