+++ runtime/src/deterministic.rs
@@ -25,8 +25,7 @@
 //! println!("Auditor state: {}", auditor.state());
 //! ```
 
-use crate::metrics::Metrics;
-use crate::{Clock, Error, Handle};
+use crate::{utils::Link, Clock, Error, Handle};
 use bytes::Bytes;
 use futures::{
     channel::mpsc,
@@ -34,7 +33,10 @@ use futures::{
     SinkExt, StreamExt,
 };
 use governor::clock::{Clock as GClock, ReasonablyRealtime};
-use prometheus_client::registry::Registry;
+use prometheus_client::{
+    metrics::{counter::Counter, family::Family},
+    registry::Registry,
+};
 use rand::{prelude::SliceRandom, rngs::StdRng, CryptoRng, RngCore, SeedableRng};
 use sha2::{Digest, Sha256};
 use std::{
@@ -52,6 +54,60 @@ use tracing::trace;
 
 /// Range of ephemeral ports assigned to dialers.
 const EPHEMERAL_PORT_RANGE: Range<u16> = 32768..61000;
+
+#[derive(Debug)]
+struct Metrics {
+    tasks_spawned: Counter,
+    task_polls: Counter,
+    bandwidth: Family<Link, Counter>,
+}
+
+impl Metrics {
+    pub fn record_task_spawned(&self) {
+        self.tasks_spawned.inc();
+    }
+
+    pub fn record_task_poll(&self) {
+        self.task_polls.inc();
+    }
+
+    pub fn record_bandwidth(&self, origin: SocketAddr, destination: SocketAddr, bytes: usize) {
+        let link = Link {
+            origin: origin.to_string(),
+            destination: destination.to_string(),
+        };
+        self.bandwidth.get_or_create(&link).inc_by(bytes as u64);
+    }
+}
+
+impl Metrics {
+    pub fn init(registry: Arc<Mutex<Registry>>) -> Self {
+        let metrics = Self {
+            bandwidth: Family::default(),
+            task_polls: Counter::default(),
+            tasks_spawned: Counter::default(),
+        };
+        {
+            let mut registry = registry.lock().unwrap();
+            registry.register(
+                "tasks_spawned",
+                "Total number of tasks spawned",
+                metrics.tasks_spawned.clone(),
+            );
+            registry.register(
+                "task_polls",
+                "Total number of task polls",
+                metrics.task_polls.clone(),
+            );
+            registry.register(
+                "bandwidth",
+                "Bandwidth usage by origin and destination",
+                metrics.bandwidth.clone(),
+            );
+        }
+        metrics
+    }
+}
 
 /// Track the state of the runtime for determinism auditing.
 pub struct Auditor {
@@ -809,5 +865,57 @@ mod tests {
                 now + Duration::new(15, 0),
             ]
         );
+    }
+
+    #[test]
+    fn test_bandwidth_metrics() {
+        let registry = Arc::new(Mutex::new(Registry::default()));
+        let metrics = Metrics::init(registry.clone());
+
+        // Send some data
+        let socket1 = SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), 8080);
+        let socket2 = SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), 8081);
+        metrics.record_bandwidth(socket1, socket2, 100);
+        metrics.record_bandwidth(socket2, socket1, 200);
+        metrics.record_bandwidth(socket2, socket1, 150);
+
+        // Verify tracking
+        assert_eq!(
+            metrics
+                .bandwidth
+                .get_or_create(&Link {
+                    origin: socket1.to_string(),
+                    destination: socket2.to_string(),
+                })
+                .get(),
+            100
+        );
+        assert_eq!(
+            metrics
+                .bandwidth
+                .get_or_create(&Link {
+                    origin: socket2.to_string(),
+                    destination: socket1.to_string(),
+                })
+                .get(),
+            350
+        );
+    }
+
+    #[test]
+    fn test_task_metrics() {
+        let registry = Arc::new(Mutex::new(Registry::default()));
+        let metrics = Metrics::init(registry.clone());
+
+        for _ in 0..5 {
+            metrics.record_task_spawned();
+        }
+
+        for _ in 0..10 {
+            metrics.record_task_poll();
+        }
+
+        assert_eq!(metrics.tasks_spawned.get(), 5);
+        assert_eq!(metrics.task_polls.get(), 10);
     }
 }
