+++ runtime/src/utils.rs
@@ -10,6 +10,7 @@ use futures::{
     stream::{AbortHandle, Abortable},
     FutureExt,
 };
+use prometheus_client::encoding::EncodeLabelSet;
 use std::{
     any::Any,
     future::Future,
@@ -111,6 +112,12 @@ where
             .poll(cx)
             .map(|res| res.map_err(|_| Error::Closed).and_then(|r| r))
     }
+}
+
+#[derive(Clone, Debug, Hash, PartialEq, Eq, EncodeLabelSet)]
+pub struct Link {
+    pub origin: String,
+    pub destination: String,
 }
 
 #[cfg(test)]
